Refine layout
