+++ IShop/Ishop/Views/Carts/AccountDetails.swift
@@ -66,23 +66,13 @@ struct AccountDetails: View {
             .padding(10)
             Divider()
             
-            Button(action: {
+            LoadingButton(title: "Continue", state: State) {
                 withAnimation {
                     State = State == .initial ? .loading : State == .loading ? .finished : .initial
                     
                 }
-            }, label: {
-                Text("Continue")
-                    .font(Font.callout.weight(.semibold))
-                    .padding()
-                    .foregroundColor(.mainBackground)
-                    .frame(maxWidth: .infinity)
-                    .background(Color.mainDark)
-                    .cornerRadius(8)
-            })
+            }
             .padding(10)
-            LoadingButton(title: "Continue", state: State)
-                .padding()
             
         }
     }
@@ -95,52 +85,3 @@ struct AccountDetails_Previews: PreviewProvider {
 }
 
 
-struct LoadingButton: View {
-    
-    enum ButtonState {
-        case initial, loading, finished
-    }
-    let title: String
-    
-    var state: ButtonState
-    @State private var animate: Bool = false
-    
-    let action: () -> ()
-    var body: some View {
-        Button(action: action) {
-            ZStack {
-                switch state {
-                case .initial:
-                    Text(title)
-                        .font(Font.callout.weight(.semibold))
-                    
-                case .loading:
-                    
-                    Circle()
-                        .trim(from: 0, to: 0.7)
-                        .stroke(Color.white, lineWidth: 2)
-                        .frame(width: 20, height: 20)
-                        .rotationEffect(Angle(degrees: animate ? 360 : 0))
-                        .animation(Animation.linear(duration: 1).repeatForever(autoreverses: false))
-                    Image(systemName: "circle.dashed")
-                        .imageScale(.large)
-                        .rotationEffect(Angle(degrees: animate ? 360 : 0))
-                        .animation(Animation.linear(duration: 1).repeatForever(autoreverses: false))
-                default :
-                    Image(systemName: "checkmark.circle")
-                        .imageScale(.large)
-                }
-            }
-            .padding()
-            .foregroundColor(.mainBackground)
-            .frame(maxWidth: .infinity)
-            .frame(height: 50)
-            .background(Color.mainDark)
-            .cornerRadius(8)
-            .animation(.easeIn)
-            .onChange(of: state) {
-                animate = $0 == .loading ? true : false
-            }
-        }
-    }
-}
